Add 'take', 'replace', 'replace_with' from std::cell::RefCell
This changeset also clarifies the language around mutable borrows.  The
message accountable_refcell paniced with when trying to take a mutable
borrow was not entirely accurate ("RefCell is already immutably borrowed"),
as panics can also occur when trying to mutably borrow a RefCell that is
already mutably borrowed.

Technically, at the time of this writing, the `take` method is not in
stable Rust, but will be in the next version (1.49) https://github.com/rust-lang/rust/issues/71395

A rust-toolchain file has been added codifying the expectation that this
crate targets stable Rust.